refactor: Rename AuthProvider context file and update imports; introduce User model
