+++ src/hooks/useGoogleAuth.ts
@@ -2,7 +2,7 @@ import { useState, useCallback } from 'react';
 import { type CredentialResponse } from "@react-oauth/google";
 import { useNavigate } from 'react-router-dom';
 import { loginWithGoogleIdToken } from '../services/authService';
-import { useAuth } from '../contexts/AuthProvider';
+import { useAuth } from '../providers/AuthProvider';
 
 interface UseGoogleAuthOptions {
   intendedDestination?: string;
