+++ src/models/user.ts
@@ -1,0 +1,6 @@
+export interface User {
+    id: string;
+    email: string;
+    pictureUrl: string;
+    fullName: string;
+}
