+++ src/routes/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthProvider';
+import { useAuth } from '../providers/AuthProvider';
 import Loading from '../components/common/Loading/Loading';
 
 interface ProtectedRouteProps {
